Adds scaffolding for Zephyr West Module
Change-type: minor
