+++ example/prj.conf
@@ -9,3 +9,6 @@ CONFIG_NEWLIB_LIBC=y
 
 # Debugging Configuration
 CONFIG_PRINTK=y
+
+# Notecard Library
+CONFIG_NOTECARD=y
