+++ zephyr/module.yml
@@ -1,0 +1,3 @@
+build:
+  cmake: .
+  kconfig: KConfig
